# kis-lab3/openapi.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "description":"Описание API для УМИС, разрабаываемого в рамках лабораторных работ по предмету КИС.",
        "title":"Umis OpenApi",
        "contact": {
            "name": "Generalov&Gladchenko",
            "email": "xsnakee@gmail.com"
        }
    },
    "servers":[
       {
          "url": "https://localhost:5000/",
          "description": "Debug",
          "variables": {
             "standName": {
                "default": "Debug"
             },
             "basePath": {
               "default": "v1"
             }
          }
       },
       {
          "url": "https://31.28.244.114:5000/",
          "description": "Development",
          "variables": {
             "standName": {
                "default": "Development"
             },
             "basePath": {
               "default": "v1"
             }
          }
       },
       {
          "url": "https://31.28.244.114:5002/",
          "description": "Demo",
          "variables": {
             "standName": {
                "default": "Demo"
             },
             "basePath": {
               "default": "v1"
             }
          }
       },
       {
          "url": "https://31.28.244.114:5004/",
          "description": "Release",
          "variables": {
             "standName": {
                "default": "Release"
             },
             "basePath": {
...
    "paths": {
...
        "/api/MediacalAnamnesis": {
            "get": {
                "tags": [ "MediacalAnamnesis" ],
                "summary": "Получить список всех медицинских анамнезов.",
                "operationId": "GetMediacalAnamnesisList",                
                "consumes": [],
                "produces": [ "application/json" ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Успешное выполнение запроса данных.",
                        "schema": { "$ref": "#/definitions/JsonResult[List[MediacalAnamnesisDto]]" }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "500": {
                        "description": "Возникло исключение на сервере.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    }
                }
            },
            "post": {
                "tags": [ "MediacalAnamnesis" ],
                "summary": "Создать новый медицинский анамнез.",
                "operationId": "CreateMediacalAnamnesis",                
                "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
                "produces": [ "application/json" ],
                "parameters": [
                    {
                        "name": "mediacalAnamnesis",
                        "in": "body",
                        "description": "Медицинский анамнез.",
                        "required": false,
                        "schema": { "$ref": "#/definitions/MediacalAnamnesisDto" }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешное создание справочника.",
                        "schema": { "$ref": "#/definitions/JsonResult[Int32]" }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "500": {
                        "description": "Возникло исключение на сервере.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    }
                }
            }
        },
        "/api/MediacalAnamnesis/{id}": {
            "get": {
                "tags": [ "MediacalAnamnesis" ],
                "summary": "Получить медицинский анамнез по идентификатору.",
                "operationId": "GetMediacalAnamnesisById",
                "consumes": [],
                "produces": [ "application/json" ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор медицинского обследования.",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное выполнение запроса данных.",
                        "schema": { "$ref": "#/definitions/JsonResult[MediacalAnamnesisDto]" }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "500": {
                        "description": "Возникло исключение на сервере.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    }
                }
            },
            "put": {
                "tags": [ "MediacalAnamnesis" ],
                "summary": "Редактировать медицинский анамнез.",
                "operationId": "EditMediacalAnamnesis",
                
                "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
                "produces": [ "application/json" ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор медицинского анамнеза.",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "mediacalAnamnesis",
                        "in": "body",
                        "description": "Медицинский анамнез.",
                        "required": false,
                        "schema": { "$ref": "#/definitions/MediacalAnamnesisDto" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное редактирование медицинского анамнеза.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "404": {
                        "description": "Медицинский анамнез с заданным идентификатором не найден.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "500": {
                        "description": "Возникло исключение на сервере",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    }
                }
            },
            "delete": {
                "tags": [ "MediacalAnamnesis" ],
                "summary": "Пометить медицинский анамнез удаленным",
                "operationId": "MediacalAnamnesisSoftDelete",
                "consumes": [],
                "produces": [ "application/json" ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор медицинского анамнеза.",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление аллергноанамнеза.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "404": {
                        "description": "Медицинский анамнез с заданным идентификатором не найден.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    },
                    "500": {
                        "description": "Возникло исключение на сервере.",
                        "schema": { "$ref": "#/definitions/JsonVoidResult" }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "PatientDto": {
            "description": "Пациент",
            "type": "object",
            "properties": {
                "surname": {
                    "description": "Имя",
                    "type": "string"
                },
                "name": {
                    "description": "Фамилия",
                    "type": "string"
                },
                "patronymic": {
                    "description": "Отчество",
                    "type": "string"
                },
                "passport": {
                    "description": "Паспорт (серия/номер)",
                    "type": "string"
                },
                "snils": {
                    "description": "СНИЛС",
                    "type": "string"
                },
                "omc": {
                    "description": "Номер ОМС",
                    "type": "string"
                },
                "phone": {
                    "description": "Список телефонов",
                    "type": "string"
                },
                "address": {
                    "description": "Адреса (прописки, проживания)",
                    "type": "string"
                },
                "email": {
                    "description": "Электронная почта",
                    "type": "string"
                },
                "birthday": {
                    "format": "date-time",
                    "description": "Дата рождения",
                    "type": "string"
                },
                "gender": {
                    "format": "int32",
                    "description": "Пол",
                    "enum": [ 0, 1, 2 ],
                    "type": "integer"
                },
                "bloodGroup": {
                    "format": "int32",
                    "description": "Группа крови",
                    "enum": [ 0, 1, 2, 3, 4, 5, 6, 7, 8 ],
                    "type": "integer"
                },
                "disabilityDescription": {
                    "$ref": "#/definitions/DisabilityDescriptionDto",
                    "description": "Уточнение/описание группы инвалидности"
                },
                "disabilityDescriptionId": {
                    "format": "int32",
                    "type": "integer"
                },
                "medicalHistories": {
                    "description": "Истории болезней",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/MedicalHistoryDto" }
                },
                "workHistory": {
                    "$ref": "#/definitions/WorkHistoryDto",
                    "description": "Трудовой анамнез"
                },
                "workHistoryId": {
                    "format": "int32",
                    "type": "integer"
                },
                "allergyAnamnesis": {
                    "$ref": "#/definitions/AllergyAnamnesisDto",
                    "description": "Аллерго анамнез"
                },
                "allergyAnamnesisId": {
                    "format": "int32",
                    "type": "integer"
                },
                "values": {
                    "description": "Значение из каталога",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/CatalogFieldValueDto" }
                },
                "createDate": {
                    "format": "date-time",
                    "description": "Дата создания объекта",
                    "type": "string"
                },
                "lastModificationDate": {
                    "format": "date-time",
                    "description": "Дата последнего изменения",
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "description": "Идентификатор сущности",
                    "type": "integer"
                }
            }
        },
        "CatalogFieldValueDto": {
            "description": "Значение поля справочника",
            "type": "object",
            "properties": {
                "catalogToCatalogField": {
                    "$ref": "#/definitions/CatalogToCatalogFieldDto",
                    "description": "Отношение справочника к полю справочника"
                },
                "catalogToCatalogFieldId": {
                    "format": "int32",
                    "type": "integer"
                },
                "value": {
                    "description": "Значение",
                    "type": "string"
                },
                "value2": {
                    "description": "Значение2 (используется для диапазонов)",
                    "type": "string"
                },
                "dianosisId": {
                    "format": "int32",
                    "type": "integer"
                },
                "mediacalAnamnesisId": {
                    "format": "int32",
                    "type": "integer"
                },
                "physiologicalIndicationId": {
                    "format": "int32",
                    "type": "integer"
                },
                "patientId": {
                    "format": "int32",
                    "type": "integer"
                },
                "medicalHistoryId": {
                    "format": "int32",
                    "type": "integer"
                },
                "workHistoryId": {
                    "format": "int32",
                    "type": "integer"
                },
                "allergyAnamnesisId": {
                    "format": "int32",
                    "type": "integer"
                },
                "medicalExaminationId": {
                    "format": "int32",
                    "type": "integer"
                },
                "disabilityDescriptionId": {
                    "format": "int32",
                    "type": "integer"
                },
                "physicalExaminationId": {
                    "format": "int32",
                    "type": "integer"
                },
                "medicalHistoryNoteId": {
                    "format": "int32",
                    "type": "integer"
                },
                "createDate": {
                    "format": "date-time",
                    "description": "Дата создания объекта",
                    "type": "string"
                },
                "lastModificationDate": {
                    "format": "date-time",
                    "description": "Дата последнего изменения",
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "description": "Идентификатор сущности",
                    "type": "integer"
                }
            }
        },
...
        "MedicalHistoryDto": {
            "description": "История болезни",
            "type": "object",
            "properties": {
                "patient": {
                    "$ref": "#/definitions/PatientDto",
                    "description": "Пациент"
                },
                "patientId": {
                    "format": "int32",
                    "type": "integer"
                },
                "medicalAnamneses": {
                    "description": "Анамнезы",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/MediacalAnamnesisDto" }
                },
                "physiologicalIndications": {
                    "description": "Физиологические показания",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/PhysiologicalIndicationDto" }
                },
                "medicalExaminations": {
                    "description": "Обследования",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/MedicalExaminationDto" }
                },
                "physicalExaminations": {
                    "description": "Осмотры",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/PhysicalExaminationDto" }
                },
                "diagnoses": {
                    "description": "Диагнозы",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/DiagnosisDto" }
                },
                "notes": {
                    "description": "Заметки",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/MedicalHistoryNoteDto" }
                },
                "assignments": {
                    "description": "Назначения",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/AssignmentDto" }
                },
                "values": {
                    "description": "Значение из каталога",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/CatalogFieldValueDto" }
                },
                "createDate": {
                    "format": "date-time",
                    "description": "Дата создания объекта",
                    "type": "string"
                },
                "lastModificationDate": {
                    "format": "date-time",
                    "description": "Дата последнего изменения",
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "description": "Идентификатор сущности",
                    "type": "integer"
                }
            }
        },
...
        "MediacalAnamnesisDto": {
            "type": "object",
            "properties": {
                "anamnesisType": {
                    "format": "int32",
                    "description": "Тип анамнеза",
                    "enum": [ 0, 1, 2, 3, 4 ],
                    "type": "integer"
                },
                "medicalHistory": {
                    "$ref": "#/definitions/MedicalHistoryDto",
                    "description": "История болезни"
                },
                "medicalHistoryId": {
                    "format": "int32",
                    "type": "integer"
                },
                "description": {
                    "description": "Словесное описание",
                    "type": "string"
                },
                "values": {
                    "description": "Значение из каталога",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/CatalogFieldValueDto" }
                },
                "createDate": {
                    "format": "date-time",
                    "description": "Дата создания объекта",
                    "type": "string"
                },
                "lastModificationDate": {
                    "format": "date-time",
                    "description": "Дата последнего изменения",
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "description": "Идентификатор сущности",
                    "type": "integer"
                }
            }
        },
        "PhysiologicalIndicationDto": {
            "description": "Физиологические показания",
            "type": "object",
            "properties": {
                "medicalHistory": {
                    "$ref": "#/definitions/MedicalHistoryDto",
                    "description": "История болезни"
                },
                "medicalHistoryId": {
                    "format": "int32",
                    "type": "integer"
                },
                "note": {
                    "description": "Словесное описание",
                    "type": "string"
                },
                "values": {
                    "description": "Значение из каталога",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/CatalogFieldValueDto" }
                },
                "createDate": {
                    "format": "date-time",
                    "description": "Дата создания объекта",
                    "type": "string"
                },
                "lastModificationDate": {
                    "format": "date-time",
                    "description": "Дата последнего изменения",
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "description": "Идентификатор сущности",
                    "type": "integer"
                }
            }
        },
...
        "MedicalHistoryNoteDto": {
            "description": "Заметка для истории болезни",
            "type": "object",
            "properties": {
                "subject": {
                    "description": "Тема",
                    "type": "string"
                },
                "content": {
                    "description": "Текст",
                    "type": "string"
                },
                "medicalHistory": {
                    "$ref": "#/definitions/MedicalHistoryDto",
                    "description": "История болезни"
                },
                "medicalHistoryId": {
                    "format": "int32",
                    "type": "integer"
                },
                "values": {
                    "description": "Значение из каталога",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/CatalogFieldValueDto" }
                },
                "createDate": {
                    "format": "date-time",
                    "description": "Дата создания объекта",
                    "type": "string"
                },
                "lastModificationDate": {
                    "format": "date-time",
                    "description": "Дата последнего изменения",
                    "type": "string"
                },
                "id": {
                    "format": "int32",
                    "description": "Идентификатор сущности",
                    "type": "integer"
                }
            }
        },
...
        "JsonVoidResult": {
            "description": "Стандартный HTTP Ответ",
            "type": "object",
            "properties": {
                "statusCode": {
                    "format": "int32",
                    "description": "Код ответа http",

                    "type": "integer"
                },
                "message": {
                    "description": "Сообщение об ошибке",
                    "type": "string"
                },
                "errors": {
                    "description": "Результаты валидации",
                    "type": "object",
                    "additionalProperties": { "type": "object" }
                }
            }
        },
        "JsonResult[Int32]": {
            "description": "стандартный Http ответ с данными",
            "type": "object",
            "properties": {
                "data": {
                    "format": "int32",
                    "description": "Данные",
                    "type": "integer"
                },
                "statusCode": {
                    "format": "int32",
                    "description": "Код ответа http",

                    "type": "integer"
                },
                "message": {
                    "description": "Сообщение об ошибке",
                    "type": "string"
                },
                "errors": {
                    "description": "Результаты валидации",
                    "type": "object",
                    "additionalProperties": { "type": "object" }
                }
            }
        },
...
        "JsonResult[List[MediacalAnamnesisDto]]": {
            "description": "стандартный Http ответ с данными",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные",
                    "uniqueItems": false,
                    "type": "array",
                    "items": { "$ref": "#/definitions/MediacalAnamnesisDto" }
                },
                "statusCode": {
                    "format": "int32",
                    "description": "Код ответа http",

                    "type": "integer"
                },
                "message": {
                    "description": "Сообщение об ошибке",
                    "type": "string"
                },
                "errors": {
                    "description": "Результаты валидации",
                    "type": "object",
                    "additionalProperties": { "type": "object" }
                }
            }
        },
        "JsonResult[MediacalAnamnesisDto]": {
            "description": "стандартный Http ответ с данными",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/MediacalAnamnesisDto",
                    "description": "Данные"
                },
                "statusCode": {
                    "format": "int32",
                    "description": "Код ответа http",

                    "type": "integer"
                },
                "message": {
                    "description": "Сообщение об ошибке",
                    "type": "string"
                },
                "errors": {
                    "description": "Результаты валидации",
                    "type": "object",
                    "additionalProperties": { "type": "object" }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "name": "Authorization",
            "in": "header",
            "type": "apiKey",
            "description": "Введите слово 'Bearer', затем пробел и JWT"
        }
    },
    "security": [ { "Bearer": [] } ],
...
}